requirements.txt
SpeechRecognition==3.10.0
transformers==4.36.2
requests==2.31.0
python-dotenv==1.0.0
